# [chore] [pkg/stanza] Simplify Converter worker loop