Avoid creating redundant maps and simplify the code

Similar CPU utilization, small improvement on memory footprint:

Before:
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/adapter
BenchmarkConverter
BenchmarkConverter/worker_count=1
BenchmarkConverter/worker_count=1-10         	       1	1667772333 ns/op	1796170832 B/op	48620116 allocs/op
BenchmarkConverter/worker_count=2
BenchmarkConverter/worker_count=2-10         	       1	1248354667 ns/op	1796157056 B/op	48620030 allocs/op
BenchmarkConverter/worker_count=4
BenchmarkConverter/worker_count=4-10         	       1	1168596542 ns/op	1796160080 B/op	48620015 allocs/op
BenchmarkConverter/worker_count=6
BenchmarkConverter/worker_count=6-10         	       1	1139376458 ns/op	1796167296 B/op	48620037 allocs/op
BenchmarkConverter/worker_count=8
BenchmarkConverter/worker_count=8-10         	       1	1197123000 ns/op	1796160552 B/op	48620023 allocs/op
PASS
```

After:
```
BenchmarkConverter/worker_count=1
BenchmarkConverter/worker_count=1-10         	       1	1358997208 ns/op	1776497920 B/op	48480094 allocs/op
BenchmarkConverter/worker_count=2
BenchmarkConverter/worker_count=2-10         	       1	1310407417 ns/op	1776498096 B/op	48480061 allocs/op
BenchmarkConverter/worker_count=4
BenchmarkConverter/worker_count=4-10         	       1	1241629750 ns/op	1776497248 B/op	48480050 allocs/op
BenchmarkConverter/worker_count=6
BenchmarkConverter/worker_count=6-10         	       1	1230347041 ns/op	1776499344 B/op	48480041 allocs/op
BenchmarkConverter/worker_count=8
BenchmarkConverter/worker_count=8-10         	       1	1138284542 ns/op	1776498912 B/op	48480020 allocs/op
```